Change transitionId type to ushort in welcome properties
Updated DaveMlsInvalidCommitWelcomeProperties to use ushort for the transitionId parameter and property, ensuring only unsigned 16-bit values are accepted for transition IDs.

+++ NetCord/Gateway/Voice/DaveMlsInvalidCommitWelcomeProperties.cs
@@ -2,8 +2,8 @@
 
 namespace NetCord.Gateway.Voice;
 
-internal class DaveMlsInvalidCommitWelcomeProperties(int transitionId)
+internal class DaveMlsInvalidCommitWelcomeProperties(ushort transitionId)
 {
     [JsonPropertyName("transition_id")]
-    public int TransitionId { get; set; } = transitionId;
+    public ushort TransitionId { get; set; } = transitionId;
 }
